Extension Manager removed (run at setup now) Github class Structuration > FilePath attempt at using temp file
